+++ src/main/kotlin/com/gacagebot/botevents/EventJoin.kt
@@ -1,7 +1,8 @@
 package com.gacagebot.botevents
 
-import com.gacagebot.constants.Constants
-import com.gacagebot.recources.values.strings.Strings
+import com.gacagebot.constants.Secrets
+import com.gacagebot.localizestrings.LocalizeString
+import com.gacagebot.localizestrings.StringId
 import net.dv8tion.jda.api.events.guild.GuildJoinEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
 
@@ -11,8 +12,8 @@ class EventJoin : ListenerAdapter() {
         super.onGuildJoin(event)
         event
             .guild
-            .getTextChannelById(Constants.CHANNEL_ID)
-            ?.sendMessage(Strings.first_join_server_message)
+            .getTextChannelById(Secrets.CHANNEL_ID.getString)
+            ?.sendMessage(LocalizeString.get(StringId.MEETINGS_FIRST_JOIN_IN_THE_SERVER.id))
     }
 
 }