+++ src/main/kotlin/com/gacagebot/constants/Secrets.kt
@@ -1,0 +1,6 @@
+package com.gacagebot.constants
+
+enum class Secrets(val getString: String) {
+    BOT_TOKEN("Paste your bot token here"),
+    CHANNEL_ID("Use if necessary")
+}