+++ CHANGELOG.md
@@ -13,8 +13,8 @@ This project adheres to the: [versionamento semântico](https://semver.org/spec/
 - Command to stop or skip a music track. (PR: [#1](https://github.com/GustavoVieiraDrk/G-Cage-Bot/pull/1)).
 - Implements a command manager design to make bot music commands easier. (PR: [#2](https://github.com/GustavoVieiraDrk/G-Cage-Bot/pull/2)).
 - Update README.md and add removed file by accidentally. (PR: [#4](https://github.com/GustavoVieiraDrk/G-Cage-Bot/pull/4)).
+- Strings.properties to use as a localize strings to show bot messages. (PR: [#5](https://github.com/GustavoVieiraDrk/G-Cage-Bot/pull/5)).
 
 ### Bugfix
 
 - Fix bot disconnect and reconnect to voice channel repeatedly. (PR: [#3](https://github.com/GustavoVieiraDrk/G-Cage-Bot/pull/3)).
--